Add handling in USB-xHCI for slow devices
There are new USB-Sticks, that return a TX_ERROR when the time between
two early requests is too short. If we get this error the endpoint
halts.
In this case trigger a reset for the endpoint, increase the delay
between requests and try again.
When the reset happens the following message appears:
"Resetting EP 0..."