Add fs_selftest support
We want to be able to use our main uboot for our selftests.
However, the dram test writes all through the memory, which might
corrupt for example optee memory. So we only add the code to the main
uboot, but don't activate it by default, to at least be able to build
with the main uboot.