Add support for armStoneMX8MP Rev 110
The new revision 1.10 of the armStoneMX8MP will use a new DRAM
B.ICDRAM.51, which will use the same timings as B.ICDRAM.30, as it is
used as a replacement. The new DRAM should be able to handle faster
timings. This will need to be evaluated, if the B.ICDRAM.30 will not be
used anymore.

Also add the following tags in NBoot for armStoneMX8MP:
have-adc
have-rs485
have-can0
have-can1

Also make entries for the following boards:
armStoneMX8MP-FERT10
armStoneMX8MP-FERT22
armStoneMX8MP-FERT24