video_link: Remove video_off variable
The inverted logic of video_off is rather confusing. At F&S, there
should only be a display visible if variable video_link is actually
set. If it is not set, which is the default, no display is activated.

Improve the video link boot messages to show when the video is off
and that this can be changed by setting the video_link variable.