Change splash.c to work similar to fsimx6* boards
With CONFIG_DM_VIDEO, the handling of the F&S logo was different. The
logo was only shown if variable splashimage was defined. The version
info was only shown if the splashimage was shown at position 0,0.

Change common/splash.c to behave like on previous boards:
If splashimage is defined, just show the given image with no text,
otherwise show F&S logo and version info.