Add optional init function board_video_init()
When using CONFIG_DM_VIDEO, the video is activated in stdio_add_devices()
in common/stdio.c. However it was difficult to add some board-specific
settings. If they were added to board_init() then the environment was not
available yet and the video_link variable could not be checked. If they
were added to board_late_init() it was too late already.

Add a new weak init function board_video_init() that is called rather
soon after the environment is available, but before the stdio devices
are set up.