Add display support for tcxd078iblma-98 on fsimx8mm
In picocoremx8mm-ubootinc.dtsi, define a video link for the
tcxd078iblma-98 display and activate CONFIG_PICOCOREMX8MM_MIPI_DSI
on all regular fsimx8mm platforms.

In fsimx8mm.c, add function board_video_init() to enable the
regulator for VLCD and the backlight. For now, the backlight does
not support PWM. Instead, the pin is configured as GPIO and can
only drive the full brightness. The logic of VLCD can be inverted
by setting variable vlcd_inverted to 1.

Add function splash_screen_prepare() that runs variable splashprepare
to load the splash image to the address given in variable splashimage.

Activate the necessary new features in fsimx8mm_defconfig:

- CONFIG_REGMAP and CONFIG_DM_RESET are needed for CONFIG_RESET_DISPMIX
- CONFIG_RESET_DISPMIX is used to reset the DISPMIX block
- CONFIG_CMD_BMP, CONFIG_BMP_16BPP, CONFIG_BMP_24BPP and
  CONFIG_VIDEO_BMP_RLE8 are needed to show BMP images and splash screens
- CONFIG_DM_VIDEO is needed to handle display support
- CONFIG_VIDEO_IMX_SEC_DSI is needed to handle MIPI DSI images
- CONFIG_SYS_WHITE_ON_BLACK to show the console as on previous boards
- CONFIG_VIDEO_XINLI_TCXD078IBLMA_98 to have the display driver
- CONFIG_SPLASH_SCREEN and CONFIG_SPLASH_SCREEN_ALIGN to show logos

In addition, the following settings are not ported to Kconfig yet and
need to be set in fsimx8mm.h:

- CONFIG_VIDOE_MXS to handle the lcdif driver mxsfb.c
- CONFIG_VIDEO_LOGO to add the F&S logo